requirements.txt
PyPDF2==3.0.1
python-docx==1.2.0
spacy==3.8.7
regex==2025.9.18